+++ examples/validate23ai.sql
@@ -1,4 +1,4 @@
--- Oracle23ai JSON-validate constrain
+-- Oracle23ai JSON-validate constraint
 
 DROP TABLE IF EXISTS object23ai;
 
@@ -25,6 +25,6 @@ ALTER TABLE object23ai ADD CONSTRAINT object23ai_ck_1 check (data is json valida
 }]');
 
 
-Insert into UWE.OBJECT23AI (OBJECT_NAME,DATA) values ('bananas','{"fruit":"banana","quantity":12,"orderdate":"2024-08-23"}');
-Insert into UWE.OBJECT23AI (OBJECT_NAME,DATA) values ('apples','{"fruit":"apples","quantity":10,"orderdate":"2024-08-23"}');
+INSERT INTO object23ai (OBJECT_NAME,DATA) values ('bananas','{"fruit":"banana","quantity":12,"orderdate":"2024-08-23"}');
+INSERT INTO object23ai (OBJECT_NAME,DATA) values ('apples','{"fruit":"apples","quantity":10,"orderdate":"2024-08-23"}');
 COMMIT;